requirements.txt
bdfparser
Pillow